Integrate Slurm (#72)
This is a rather large change, using PSI/J for Slurm integration. No other changes should be made until this is merged, to avoid terrible merge conflicts.

Note that almost everything has been moved out of context files into the Jsonnet config. The reason is that changing config will not invalidate the redun cache, whereas changing context will.

All Slurm jobs are sized in the Jsonnet config. To see the effect of this, from jsonnet config/eri-executor.jsonnet. Jsonnet is used to reduce the duplication in the config.

After this is merged, a following tidy-up will be to refactor all the individual tool wrappers to be native redun tasks. This will be a considerable simplification, but has been held back for now to avoid building in a dependency on redun. Which we are now happy to do.

I have run though the whole pipeline, at least in stages, and while this PR is being reviewed will do a final --no-cache run, to check everything is OK. 😅

Fixes #49
---------

Co-authored-by: Simon Guest <simon.guest@agresearch.co.nz>