Update to latest redun 0.27 with Python 3.12 and nixpkgs-unstable (#149)
Python package fancycompleter previously blocked use of Python 3.12 but someone kindly fixed that.

Switching to up-to-date nixpkgs-unstable is a nice simplification, as packages which were previously unavailable in the branch we were using (multiqc) are now readily available, and the patch for a broken release of hatch is no longer required.

The way redun is integrated has been radically simplified. The redun flake now simply exports a standard Python package which is compatible with python3.withPackages. That flake has an up-to-date redun with older textual 1.0.0 to avoid a redun console bug for now.

Consuming the gquery flake has been simplified, since that no longer exports two different packages.

Simplified packaging of R scripts.

Some other renamings to increase clarity.

Fixes #145