KGD optional plots (#151)
Run 240621_A01439_0276_AH33J5DRX5 has KGD.stdout with:
Multiple plates - plate plots not produced

I spoke with Ken. This may be simply an anomaly introduced by a change made by lab staff. So we decided not to investigate further.

This change makes this case non-fatal for the overall gbs_prism run.

Also, the complete list of missing files is included in the ClusterExecutorError now, instead of just the first.