Support for GBS uber keyfiles (#169)
Add support for GBS user keyfiles, with these two main features:

Enable GBS prism tasks to be invoked from outside of this repo
Support multiple parts in demultiplex with tassel, via ramify_tassel_keyfile, with merging of multiple parts as in the legacy gbs_prism pipeline.
External Invocation
This involves structuring the Jsonnet cluster configuration to support a library path, with the tool configuration returned from a Jsonnet function parameterised by job prefix and executor name (slurm). Further parameters could be exposed, such as the Slurm partition, but that hasn't been done at this stage.

Further, the KGD packaging was modified to explicitly export the R packages required by KGD, so these could be installed in R. Some KGD plots were made optional as there are conditions with uber keyfiles when they are not generated, which is not an error.

The passing of environment variables from gquery through gbs_prism and out to gbs-uber-keyfiles has been simplified, affecting the install script and the Nix flake.

Tassel Demultiplex
An aggregate demultiplex task is introduced, which encapsulates the various Tassel3 stages. The existing Tassel tasks have been modified to remove references to cohorts in favour of enzymes and prefixes, to make suitable for external use.

Fixes #142