Make get_unblind_script task uncacheable (#218)
We can't know ahead of time whether we need a new unblind script, so
we always fetch it, and rely on redun file comparison to not trigger
downstream processing if it hasn't in fact changed.

For this to work, we need to reset the file mtime if unblind script is unchanged,
otherwise redun thinks it's different, and all the downstream tasks get retriggered.

Fixes #195

---------

Co-authored-by: Simon Guest <simon.guest@agresearch.co.nz>